% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_method.R
\name{dfs_method}
\alias{dfs_method}
\title{dfs_method function}
\usage{
dfs_method(GraphMade = NA)
}
\arguments{
\item{GraphMade}{an existing maze graph object. Value is set to NA}
}
\value{
The function returns a maze built according to dfs algorithm with paths anf walls displayed.
}
\description{
Generate a maze using a randomized adaptation of the depth-first search (DFS) algorithm
}
\examples{
Example_maze <- myGraph(nrows = 5,ncols = 9) 
Example_maze <- dfs_method(GraphMade = Example_maze)
plotmyMaze(Example_maze,nrows= 5,ncols = 5)

}
\author{
Joost and Edward
}
