% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGraph.R
\name{myGraph}
\alias{myGraph}
\title{myGraph function}
\usage{
myGraph(nrows = 0, ncols = 0)
}
\arguments{
\item{nrows}{number of rows, maze height value arbitrarly set to 0.}

\item{ncols}{number of cols, maze weight value arbitrarly set to 0.}
}
\value{
The function returns a undirected graph network with nodes and edges according to the nrows and ncols defined
}
\description{
myGraph function
}
\details{
The function creates an undirected network with edges and nodes. 
These elements are translated to 'walls' are ON to build the maze.
}
\examples{
Example_maze <- myGraph (nrows = 5,ncols = 9) 
Example_maze <- dfs_method (GraphMade = Example_maze)
plotmyMaze(Example_maze,nrows= 5,ncols = 5)
}
\author{
Joost and Edward
}
