% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solverGraph.R
\name{solverGraph}
\alias{solverGraph}
\title{solverGraph function}
\usage{
solverGraph(GraphMade = NA, nrows = 0, ncols = 0)
}
\arguments{
\item{GraphMade}{An existing graph}

\item{nrows}{number of rows, maze height value arbitrarly set to 0.}

\item{ncols}{number of cols, maze weight value arbitrarly set to 0.}
}
\value{
The function returns returns a green path indicating the solution of the maze.
}
\description{
solverGraph function
}
\details{
The function transforms the graph into a data frame and then 
creates a plot of the existing maze with ggplot2 package.
}
\examples{
Example_maze <- myGraph (nrows = 5,ncols = 9) 
Example_maze <- dfs_method (GraphMade = Example_maze)
solverGraph(Example_maze,nrows= 5,ncols = 5)
}
\author{
Joost and Edward
}
